+++ smd/g_spawn.c
@@ -496,7 +496,7 @@ char *ED_NewString (char *string)
 	char	*newb, *new_p;
 	int		i,l;
 	
-	l = strlen(string) + 1;
+	l = (int)strlen(string) + 1;
 
 	newb = gi.TagMalloc (l, TAG_LEVEL);
 
@@ -616,7 +616,7 @@ char *ReadTextFile (char *filename, int *size)
 		return NULL;
     }
 
-    len = fread(filestring, 1, len, fp);
+    len = (int)fread(filestring, 1, len, fp);
 	*size = len;
     filestring[len] = 0;
 
@@ -668,7 +668,7 @@ qboolean LoadAliasFile (char *name)
 			Com_sprintf(pakfile, sizeof(pakfile), "%s/pak%d.pak",filename,i);
 			if (NULL != (fpak = fopen(pakfile, "rb")))
 			{
-				num=fread(&pakheader,1,sizeof(pak_header_t),fpak);
+				num=(int)fread(&pakheader,1,sizeof(pak_header_t),fpak);
 				if(num >= sizeof(pak_header_t))
 				{
 					if (pakheader.id[0] == 'P' &&
@@ -691,7 +691,7 @@ qboolean LoadAliasFile (char *name)
 									gi.dprintf("LoadAliasData: Memory allocation failure for entalias.dat\n");
 									return false;
 								}
-								alias_data_size = fread(alias_data,1,pakitem.size,fpak);
+								alias_data_size = (int)fread(alias_data,1,pakitem.size,fpak);
 								alias_data[pakitem.size] = 0; // put end marker
 								alias_from_pak = true;
 							}
